# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Animal Rescue India API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /volunteers:
    get:
      operationId: getVolunteers
      tags: [volunteers]
      summary: Get registered volunteers
      parameters:
        - name: district
          in: query
          schema:
            type: string
        - name: state
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of volunteers
          content:
            application/json:
              schema:
                type: object
                properties:
                  volunteers:
                    type: array
                    items:
                      $ref: "#/components/schemas/Volunteer"
                  total:
                    type: number
                required: [volunteers, total]
    post:
      operationId: registerVolunteer
      tags: [volunteers]
      summary: Register as volunteer/NGO
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateVolunteerRequest"
      responses:
        "201":
          description: Volunteer registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Volunteer"
  /vets:
    get:
      operationId: getVets
      tags: [vets]
      summary: Get veterinary doctors
      parameters:
        - name: district
          in: query
          schema:
            type: string
        - name: state
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of vets
          content:
            application/json:
              schema:
                type: object
                properties:
                  vets:
                    type: array
                    items:
                      $ref: "#/components/schemas/Vet"
                  total:
                    type: number
                required: [vets, total]
    post:
      operationId: createVet
      tags: [vets]
      summary: Add a veterinarian
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateVetRequest"
      responses:
        "201":
          description: Vet added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vet"
  /hospitals:
    get:
      operationId: getHospitals
      tags: [vets]
      summary: Get veterinary hospitals
      parameters:
        - name: district
          in: query
          schema:
            type: string
        - name: state
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of hospitals
          content:
            application/json:
              schema:
                type: object
                properties:
                  hospitals:
                    type: array
                    items:
                      $ref: "#/components/schemas/Hospital"
                  total:
                    type: number
                required: [hospitals, total]
    post:
      operationId: createHospital
      tags: [vets]
      summary: Add a hospital
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateHospitalRequest"
      responses:
        "201":
          description: Hospital added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Hospital"
...
components:
  schemas:
...
    Volunteer:
      type: object
      properties:
        id:
          type: number
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        type:
          type: string
        organizationName:
          type: string
          nullable: true
        district:
          type: string
        state:
          type: string
        skills:
          type: string
          nullable: true
        availability:
          type: string
          nullable: true
        vehicleAvailable:
          type: boolean
        status:
          type: string
        createdAt:
          type: string
      required:
        - id
        - name
        - email
        - phone
        - type
        - district
        - state
        - vehicleAvailable
        - status
        - createdAt
    CreateVolunteerRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        type:
          type: string
        organizationName:
          type: string
          nullable: true
        district:
          type: string
        state:
          type: string
        skills:
          type: string
          nullable: true
        availability:
          type: string
          nullable: true
        vehicleAvailable:
          type: boolean
      required:
        - name
        - email
        - phone
        - type
        - district
        - state
        - vehicleAvailable
    Vet:
      type: object
      properties:
        id:
          type: number
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        specialization:
          type: string
          nullable: true
        registrationNumber:
          type: string
        hospitalId:
          type: number
          nullable: true
        hospitalName:
          type: string
          nullable: true
        district:
          type: string
        state:
          type: string
        isAvailable:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - name
        - email
        - phone
        - registrationNumber
        - district
        - state
        - isAvailable
        - createdAt
    CreateVetRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        specialization:
          type: string
          nullable: true
        registrationNumber:
          type: string
        hospitalId:
          type: number
          nullable: true
        district:
          type: string
        state:
          type: string
        isAvailable:
          type: boolean
      required:
        - name
        - email
        - phone
        - registrationNumber
        - district
        - state
        - isAvailable
    Hospital:
      type: object
      properties:
        id:
          type: number
        name:
          type: string
        address:
          type: string
        district:
          type: string
        state:
          type: string
        phone:
          type: string
        email:
          type: string
          nullable: true
        latitude:
          type: number
          nullable: true
        longitude:
          type: number
          nullable: true
        totalBeds:
          type: number
          nullable: true
        availableBeds:
          type: number
          nullable: true
        isGovernment:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - name
        - address
        - district
        - state
        - phone
        - isGovernment
        - createdAt
    CreateHospitalRequest:
      type: object
      properties:
        name:
          type: string
        address:
          type: string
        district:
          type: string
        state:
          type: string
        phone:
          type: string
        email:
          type: string
          nullable: true
        latitude:
          type: number
          nullable: true
        longitude:
          type: number
          nullable: true
        totalBeds:
          type: number
          nullable: true
        availableBeds:
          type: number
          nullable: true
        isGovernment:
          type: boolean
      required:
        - name
        - address
        - district
        - state
        - phone
        - isGovernment
...